all changes**
Up until now, I have made the changes in  TI using lab4 torque control, which uses Field Oriented Control Solution. TI codes ae built for a constant DC voltage in the board across the DC bus. However, for the WindCar, the board will control the motor side resulting in controlling th ebus voltage , resulting in maximum power extraction from the wind generator for a given wind speed. The lab was modified such that we can read the dc bus current, we can define a reference for the q current of the motor side so that creating a net delta i going into the input capacitor will make  the desired voltage change. After adjusting the input voltage, the iq will again followthe generator side current. The experiments were conducted using a variable DC voltage, with hand tuning for the dc voltage imitating the change on it over time to observe the effect on the Iq  ref current. This was achieved within the code.

However, we realized that we don't have the confidence over the current and the control variables. So it was decided to monitor the current, voltage and the reference Id-q values for the operation.

For this, the lab1b-c codes are used. These codes use park-clarke transformations as well but apply V/f contol. The graph plotting of these codes were extracted and integrated to lab4. The effects on lab4 will beinvestigated further.

In the code plots and for the expressions Q 24 value is selected. However, this creates a conversion factor between the real value observed with the oscilloscope and the one on the TI real-time data. This will be examined.