Plotting and windspeed integration back again
The plots for the phase current, phase voltage,reference Iq and measured Iq are successfully integrated to the lab4, (in txt files). The csv files have the graphs showing for a reference change in Iq, the real measured data follows the reference with some noise. It was noted that the measured value is nearly constant and meaningful when there is a load torque on the shaft. Otherwise, the Iq_measured has an oscillating beahviour (we think due to the lack of oad torque and limited speed max, the torque is decreased but the torque control then increases its value again and this creates an oscillation overall Te-TL=dw/dt, dw/dt is limited, and goes 0 then TL=0, Te=0 ...)

moreover, the previous changes for the wind simulating behaviour for Vdc_ref-windspeed related Iq_ref determination is inserted and the results will be showed tomorrow for further verification.